// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/KnowledgeTreeNodeDto.ts
export interface KnowledgeTreeNodeDto {
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    parentId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    stageId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    gradeId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    subjectId: string;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    nodeCode: string;
    /**
     * 对应数据库 knowledge_node.name。
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    name: string;
    /**
     * 开放业务代码；当前未冻结枚举。
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    nodeType: string;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeTreeNodeDto
     */
    readonly levelNo: number;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeTreeNodeDto
     */
    difficulty?: number | null;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeTreeNodeDto
     */
    keywords?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof KnowledgeTreeNodeDto
     */
    enabled: boolean;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeTreeNodeDto
     */
    sortOrder: number;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeTreeNodeDto
     */
    readonly version: number;
    /**
     * 
     * @type {Date}
     * @memberof KnowledgeTreeNodeDto
     */
    readonly createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof KnowledgeTreeNodeDto
     */
    readonly updatedAt: Date;
    /**
     * 
     * @type {Array<KnowledgeTreeNodeDto>}
     * @memberof KnowledgeTreeNodeDto
     */
    children: Array<KnowledgeTreeNodeDto>;
}