// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ScorePageResponseAllOfData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ScoreListItemDto } from './ScoreListItemDto';
import {
    ScoreListItemDtoFromJSON,
    ScoreListItemDtoFromJSONTyped,
    ScoreListItemDtoToJSON,
    ScoreListItemDtoToJSONTyped,
} from './ScoreListItemDto';

/**
 * 
 * @export
 * @interface ScorePageResponseAllOfData
 */
export interface ScorePageResponseAllOfData {
    /**
     * 
     * @type {number}
     * @memberof ScorePageResponseAllOfData
     */
    page: number;
    /**
     * 
     * @type {number}
     * @memberof ScorePageResponseAllOfData
     */
    pageSize: number;
    /**
     * 
     * @type {number}
     * @memberof ScorePageResponseAllOfData
     */
    total: number;
    /**
     * 
     * @type {number}
     * @memberof ScorePageResponseAllOfData
     */
    totalPages: number;
    /**
     * 
     * @type {Array<ScoreListItemDto>}
     * @memberof ScorePageResponseAllOfData
     */
    items?: Array<ScoreListItemDto>;
}

/**
 * Check if a given object implements the ScorePageResponseAllOfData interface.
 */
export function instanceOfScorePageResponseAllOfData(value: object): value is ScorePageResponseAllOfData {
    if (!('page' in value) || value['page'] === undefined) return false;
    if (!('pageSize' in value) || value['pageSize'] === undefined) return false;
    if (!('total' in value) || value['total'] === undefined) return false;
    if (!('totalPages' in value) || value['totalPages'] === undefined) return false;
    return true;
}

export function ScorePageResponseAllOfDataFromJSON(json: any): ScorePageResponseAllOfData {
    return ScorePageResponseAllOfDataFromJSONTyped(json, false);
}

export function ScorePageResponseAllOfDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScorePageResponseAllOfData {
    if (json == null) {
        return json;
    }
    return {
        
        'page': json['page'],
        'pageSize': json['pageSize'],
        'total': json['total'],
        'totalPages': json['totalPages'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(ScoreListItemDtoFromJSON)),
    };
}

export function ScorePageResponseAllOfDataToJSON(json: any): ScorePageResponseAllOfData {
    return ScorePageResponseAllOfDataToJSONTyped(json, false);
}

export function ScorePageResponseAllOfDataToJSONTyped(value?: ScorePageResponseAllOfData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'page': value['page'],
        'pageSize': value['pageSize'],
        'total': value['total'],
        'totalPages': value['totalPages'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(ScoreListItemDtoToJSON)),
    };
}