// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/SubjectDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SubjectDto
 */
export interface SubjectDto {
    /**
     * 
     * @type {string}
     * @memberof SubjectDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SubjectDto
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof SubjectDto
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof SubjectDto
     */
    enabled: boolean;
}

/**
 * Check if a given object implements the SubjectDto interface.
 */
export function instanceOfSubjectDto(value: object): value is SubjectDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    return true;
}

export function SubjectDtoFromJSON(json: any): SubjectDto {
    return SubjectDtoFromJSONTyped(json, false);
}

export function SubjectDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubjectDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'code': json['code'],
        'name': json['name'],
        'enabled': json['enabled'],
    };
}

export function SubjectDtoToJSON(json: any): SubjectDto {
    return SubjectDtoToJSONTyped(json, false);
}

export function SubjectDtoToJSONTyped(value?: SubjectDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'code': value['code'],
        'name': value['name'],
        'enabled': value['enabled'],
    };
}